refactor(os/interfaces): split common.py into sessions.py + streaming.py
The codebase convention is domain-named files (agent.py, audio.py,
router.py, auth.py). Generic fallback names like common.py / utils.py
only show up where no domain fits. common.py spanned two unrelated
domains (session store + streaming event handling), which is why no
single-word name felt right.

Split into two domain-named files:

- libs/agno/agno/os/interfaces/sessions.py — SessionStoreConfig,
  build_session_store_config, find_latest_session_id
  (everything related to finding / creating DB sessions for an entity)

- libs/agno/agno/os/interfaces/streaming.py — normalize_event,
  SUPPRESSED_IN_WORKFLOW, member_name, task_id, collect_media_from_chunk,
  chunk_text (everything related to handling agent streaming events and
  chunking output to fit platform message limits)

is_dev_mode was a one-line os.getenv check with a single caller
(discord/security.py). Inlined there; common.py deleted.

Call sites updated:
  discord/events.py   → imports from streaming
  discord/router.py   → imports from sessions + streaming
  discord/state.py    → imports from sessions + streaming
  discord/security.py → inlined is_dev_mode
  whatsapp/router.py  → imports from sessions

Each import is now self-explanatory: you see "from ...sessions import ..."
and know exactly what you're getting.
