+++ libs/agno/agno/os/interfaces/discord/router.py
@@ -50,7 +50,8 @@ from agno.os.interfaces.discord.state import (
     _build_session_scope,
     insert_sentinel_session,
 )
-from agno.os.interfaces.common import build_session_store_config, chunk_text, find_latest_session_id
+from agno.os.interfaces.sessions import build_session_store_config, find_latest_session_id
+from agno.os.interfaces.streaming import chunk_text
 from agno.run.agent import RunOutput
 from agno.run.team import TeamRunOutput
 from agno.team import RemoteTeam, Team
