+++ libs/agno/agno/os/interfaces/discord/security.py
@@ -2,7 +2,6 @@ import os
 import time
 from typing import Any, Optional
 
-from agno.os.interfaces.common import is_dev_mode
 from agno.utils.log import log_warning
 
 # PyNaCl is optional — resolve BadSignatureError at import time so except
@@ -45,7 +44,7 @@ def verify_discord_signature(
     *,
     public_key: Optional[str] = None,
 ) -> bool:
-    if is_dev_mode():
+    if os.getenv("APP_ENV", "").lower() == "development":
         log_warning("Bypassing Discord signature validation in development mode")
         return True
